+++ src/pages/auth/SignupPage.tsx
@@ -122,7 +122,10 @@ export function SignupPage() {
       await session.storage.addHistorySpaceCreated({ user: session.user })
 
       // Add a "welcome" credential to storage
-      await session.storage.addCredential({ credential: welcomeCredential })
+      await session.storage.addCredential({
+        credential: welcomeCredential,
+        user: session.user
+      })
       login(session)
       navigate('/dashboard')
     } catch (err) {